You can rename these files (and the restart file names as desired) or you can run this as is. As is, this pair runs with
% gx kh01
% gx kh01a

This is a secondary instability calculation (for benchmarking).
The frequency should be as in the kh01a output file in the git repo, ie,
w = (0, 4.9).

The ky = 0.1 mode is set up in the kh01 run. It is fixed in time in the
kh01a case. The other modes should all grow with the same growth rate, which (according
to Rogers, et al., PRL 2000, on ITG modes) should be

gamma = kx ky Phi.

In our normalization, the value of Phi is doubled, to account for ky < 0 modes. Thus we should
get, in our units,

gamma ~ 2 kx ky scale = 2 / 200 * 500 = 5.

We actually get 4.9 because Phi = 494. The difference from 500 comes from solving quasineutrality
for the (kx, ky) = (0, 0.1) mode, with the *density* set to 500. Getting 494 is the correct
FLR-corrected potential, for this value of k_perp.

The GS2 normalizations are different so the result is different (in that code's units).