I had commented out a line in the collision operator (for m=0) and that line should not have been commented out.
There is still a problem that I don't understand, related to collisions. I am trying to reproduce the growth rates from
Plamen Ivanov's paper on Z-pinch physics in the collisional limit. There are problems at long wavelength, when collisions
are turned on. I have been comparing with GS2. The results are good without collisions. But with...the errors are pretty big.

In particular, I have been setting the ion temperature to values other than unity. I will circle back to see if the
errors disappear with collisions on, but with temp = 1. I have been using non-unity temp simply as a consequence of how
I happen to track down the error that this check-in fixes.