Changes in timestep are not complete for SDCe and the new SPP method.
Mostly the rest of the changes are minor. There is a bug fix in the field solver,
related to the jacobian. And there is a bug fix in the collision operator in
linear.cu (error was in the m=0 equations).

Added a crude diagnostic for primary, secondary and tertiary amplitudes.

Began to add the capability to set the CUDA device id at runtime. Still hard-wired
to devid = 0 for now.

There were problems in the various timesteppers for the cases with eqfix = T. All
of these problems are corrected, except in the RK3 method.