Attempting to improve performance with better management of some of the kernel launching and indexing.
I tested this code for a well-resolved slab ETG nonlinear case, as well as kh01/kh01a, etc.

There are no obvious bugs associated with these changes, but I am checking this in now because
I have to work on something else and I don't want to lose the progress.

Also, this work was done on Traverse for a change. No problems with that, but I was fiddling
with the devid parameter. It should be pointing to devid = 0, which is appropriate for
gx.umd.edu. I want to promote that variable to the command line, but that is tied up
with the move to multi-GPU operation, so I am waiting for now.