Mostly working on minor optimizations, readability, and allowing for multiple species
With respect to multispecies runs, there were a few things that were obviously incomplete.

The closures had no factors of v_t or T/Z.

The quasineutrality equation had missing factors of n*Z.

The CFL condition is now calculated based on the non-gyroaverage Phi, instead of
the gyroaveraged Phi from the last species in the list of species (which was just
the result of the single-species coding, not an intentional choice). Using Phi
is always the more conservative choice, but not by very much for typical input files.

With twist-and-shift boundary conditions, the free streaming term was not calculated
for more than the first kinetic species. The coding for free streaming was correct
for periodic boundary conditions.

I believe there were additional missing normalizations here and there. I have not
yet run the final multispecies tests. Coming soon.

There is still no ability to run without an adiabatic species in this main branch.

I moved hypercollisions to the end of the time advance sequence. I put in some new
functions (JT, JU) to reduce multiple instances of repeated and tedious coding.