There was a problem that I didn't catch with the last update, for a case with collisions. This is resolved here.
There was a bug that has been around for a while, associated
with closures. The closure option variable was not initialized to zero
which meant that the behavior was not guaranteed for cases which
should not have had closures -- but this would have caused a segfault.
So this was not a bug that had had an effect.

I moved the reality enforcement routines out of the grad_parallel
operator and into run_gx, to be called every nreal timesteps. nreal is set
to default to 10000. This is just to keep floating point errors from accumulating
in a way that is non-physically wrong.