Removed need for forward dependencies and cleaned up warnings tracked down with valgrind.
There was a problem with the routines that read from eik.out files. I changed the logic and I will
check in some tests for them shortly. They are not failing but I have not yet tried to be sure they are
fully correct and general.

The standard test cases give the same answers with these changes. At least two intermittent memory
errors seem to be improved.