Added git version information, the compilation date, the compilation time, the name of the user who performed the compilation, and the hostname of the system where the compilation was done to the NetCDF output.
At each invocation of make, this information is compiled as version.c. After the executable is built, version.c
is moved to the old/ directory. It could have been removed altogether, but that would be pretty confusing.
Moving it forces the dependency rule in the Makefile to rebuild version.c. Do not put version.c into
revision control.

There is a header file in the include directory called include/version.h. This file holds the information so that
it is available for writing.