First installment of work to allow restarts with changes to the system size (box size, resolution, etc).
What is here seems to be working correctly.

I have not yet made all combinations possible. The most notable missing items are the ability to change the
resolution or box size in the z-direction, and the ability to reduce the grid resolution in x or y.

Increasing resolution in x, y work. Change the perpendicular box sizes works (in the absence of magnetic shear).

Changing velocity space resolution works.