Adding some stellarator geometry files from Rogerio Jorge.
Documentation for these files is not yet provided, but I have
tested for correctness using python3 and

pip3 install qsc

to pick up the required QSC library from Matt Landreman and his
colleagues.

The key file here is <geo/NearAxis_GS2_GX.py>. This file creates
a geometry file for a quasisymmetric stellarator configuration. The
file that is created is called <gxinput.out>. It uses the format
of <eik.out> from the gs2 suite.

To read a geometry like this, you must specify ntheta to match
the values used in the python script (which you are free to edit).

I am checking this in today for safe keeping. Once it is fully
benchmarked and test cases are available, I will go back and
provide proper documentation and ancillary files such as are
needed for various applications.