At some point in the recent past, a bug introduced for the case of Beer 4+2 closures with magnetic shear.
It did not matter for the sequence of ETG calculations completed through this year,
because it only mattered when magnetic shear is non-zero.

It started to show up when we began to examing stellarator cases. Formally, many
of the cases have had shat = 0, since we were using the near-axis expansion. Thus,
the problem was not seen. However, there were inconsistencies with the way the
coordinate system was defined for stellarators that did show up, and these became
conflated with the present problem. When we got the near-axis cases fixed,
I thought that was the whole problem, but of course there are also stellarator
cases with finite magnetic shear, and these were buggy because of this problem. e

The error was simply that the cufft plan for inverse parallel transforms was
commented out instead of created and so the d/dz terms with linked boundary
conditions were not correct.