Bug fixed!
Over the summer I added a NetCDF interface to the stellarator geometry specification.
That interface had a bug, which is fixed here.

The curvature and grad-B drifts used by GX are different from those used by GS2
because of the change in the definition of vt. The ASCII interface (igeo=1) was fine
but the binary interface (igeo=2) did not have the required conversion factor.