Input and output files for the L=8, M=16 case in Fig 3 of Mandell, et al., 2018.
The agreement is not exact because of ancillary changes to the model equations, related primarily (I believe) to details
of the FLR model, and specifically to the treatment of the highest Laguerre moment in the FLR terms.

The last time we worked through these benchmarks, I saved this case as inputs/cyclone_linear/cyc_lin and this file,
cyc02, reproduces that one. At that time, July 2021, there were further small changes to the FLR model that brought the
results fully into line with the published figure, but I do not have those further changes handy today. I will get back to it.