Hegna model implemented in the sense of equations for density, parallel flow, parallel and perpendicular temperatures, and parallel flux of parallel and perpendicular energy, (so six moments total), with the expectation that one will select large values of vnewk to obtain the fluid limit.
In units of the normalizing equilibrium length scale, the mean free path
for a single-species simulation is 1 / vnewk. Presumably the mean free path should
be small compared to the parallel wavelength of the instability eigenmode, which is
often of order the connection length. Assuming the equilibrium normalizing length is
the major radius, one should then presumably set mfp ~ 1/vnewk << 2 pi q,
or vnewk >> 1/(2 pi q). A few initial attempts suggest vnewk = 1.0 for the ion vnewk
parameter is in this limit for ITG modes.

If the value of vnewk is too large, then the threshold will basically be the fluid
threshold, which can be ~ 2x the kinetic threshold. This may or may not be problematic
for any given situation.

Note that for large values of vnewk, T_parallel ~ T_perp, so that there is effectively
only a total temperature perturbation (ie, there is strong pitch-angle scattering).

The model here does not require k_perp rho << 1. In fact, the dissipative terms
essentially vanish at long wavelength, so it is better to include smaller perpendicular
scales in a nonlinear simulation, so that there is sufficient damping provided by the
collisional heat fluxes.

A more direct implementation of the Hegna model would include phenomenological damping
in the density and parallel momentum equations.