Improved the geometry submodule by allowing the user to prepend an output path for the geometry files (ASCII and NetCDF). The new variable is called "out_path" and defaults to "./", so that it should be exactly backwards compatible.
There is also a vmec_path variable in the namelist, to allow for
the vmec data (the wout file) to be in some directory relative
to the execution directory. This also defaults to "./" so that
it is backward compatible.

If anyone has a change in their workflows that is a result of a
failure of our attempt to maintain backward compatibility, please
let me (Bill) know quickly.