create SpectraCalc (virtual) base class and lots of derived spectra calculator classes which compute a particular type of spectra (e.g. SpectraCalc_kxst computes spectra of the form f(kx, s, t)). these classes use the new generalized Reduction object.
the base class (SpectraCalc) also has a method for defining the netcdf variable corresponding to a particular spectra (for a particular diagnostic) and a method for writing the data to netcdf.

also creating a packager class AllSpectra (completely defined in spectra_calc.h) which stores a copy of all the possible types of spectra calculators, allowing all the calculators to be easily passed to other objects.