We had some extra factors of 2 here and there which are removed in this update. The only thing that will change is that the values stored for the magnetic drifts in output files will not be consistent after this update. The values in the output files will be increased by a factor of two relative to what we have saved in the past. We should regenerate the test cases as a result, to avoid questions later. The old code was not incorrect, so this is not a bug fix -- this is simply clean-up. Now, when you look at the outboard midplane of an up-down symmetric, axisymmetric equilibrium and inspect the value of the curvature drift factor there, it will be 1/R instead of 1/2R.
===============================

Details:

In device_funcs, we multiplied cvdrift, cvdrift0, gbdrift, and gbdrift0 by 2 when we used them.

In geometry, we defined these quantities with an extra factor of 2 in the denominator relative to
straightforward definitions.

When we read in geometric data, we were dividing these magnetic drift terms by 4. Now, they
are divided by a factor of 2. These old geometric quantity conventions are a factor of 2
larger than the straightforward definitions. So when we read them in, we are just
compensating for this old "unconventional convention".